add BeatRhythmToVisibleConverter and make TGridUnitLine only show first beat when editor is lock

+++ OngekiFumenEditor/Modules/FumenVisualEditor/ViewModels/TGridUnitLineViewModel.cs
@@ -6,7 +6,7 @@ namespace OngekiFumenEditor.Modules.FumenVisualEditor.ViewModels
     {
         public double Y { get; set; }
         public TGrid TGrid { get; set; }
-        public double BeatRhythm { get; set; }
+        public int BeatRhythm { get; set; }
         public override string ToString() => $"{Y:F4} {TGrid}";
     }
